requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.1.5
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
executing==1.2.0
factory-boy==3.2.1
Faker==18.9.0
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
